requirements.txt
numpy==1.19.4
Pillow==8.0.1
ffmpeg-python==0.2.0